Tests/LibWeb/Text/expected/HTML/parser-streams-named-character-reference-split.txt

length: 1
code point: 8713
